Updated Index.html to have a rough submit button.
Essentially, added a raw and simple HTML "Submit" button. It's not much but it's a start for our project! We would need to make a layout for the front page later, but our main concern should have the functionality of uploading a file and then having javascript send it to our database on Azure or Heroku. After that, have our website play the video via another javascript call unto the HTML page.